Add files via upload
This is a script i created to automatically create snort rules based on a document with specified indicators of compromise (IOCs). If the IOC file list is coming from a windows environment you will need to run the command "tr -d '\r' < $winIOCFile > $NewIOCFile" in order to convert the LF/CR to a linux formatted document.